+++ test/election.test.ts
@@ -1,34 +1,34 @@
 import { expect } from 'chai';
 import * as sinon from 'sinon';
 
-import { Election, Etcd3 } from '../src';
-import { getOptions, tearDownTestClient } from './util';
+import { BehaviorSubject, Observable } from 'rxjs';
+import { filter, first, share, take, takeWhile } from 'rxjs/operators';
+import { CampaignState, Election, Etcd3, IKeyValue, InvalidOperationError } from '../src';
+import { delay } from '../src/util';
+import { createTestClient, getOptions, tearDownTestClient } from './util';
 
-const sleep = (t: number) => new Promise(resolve => setTimeout(resolve, t));
-
-describe('election', () => {
+describe('elect()', () => {
   let client: Etcd3;
   let election: Election;
 
   beforeEach(async () => {
-    client = new Etcd3(getOptions());
-    election = new Election(client, 'test-election', 1);
+    client = createTestClient();
+    election = client.election('test-election', 1);
     await election.campaign('candidate');
   });
 
   afterEach(async () => {
-    election.removeAllListeners();
     await election.resign();
     await tearDownTestClient(client);
   });
 
   describe('campaign', () => {
     it('should wait for elected in campaign', async () => {
-      const client2 = new Etcd3(getOptions());
-      const election2 = new Election(client2, 'test-election', 1);
+      const client2 = createTestClient();
+      const election2 = client2.election('test-election', 1);
 
-      const client3 = new Etcd3(getOptions());
-      const election3 = new Election(client3, 'test-election', 1);
+      const client3 = createTestClient();
+      const election3 = client3.election('test-election', 1);
 
       /**
        * phase 0: client elected
@@ -42,72 +42,115 @@ describe('election', () => {
         .then(() => election.getLeader())
         .then(currentLeaderKey => {
           expect(phase).to.equal(1);
-          expect(currentLeaderKey).to.equal(election2.leaderKey);
+          expect(currentLeaderKey).to.equal(election2.campaignData!.leaderKey);
         });
 
       // essure client2 has joined campaign before client3
-      await sleep(100);
+      await delay(100);
 
       const waitElection3 = election3
         .campaign('candidate3')
         .then(() => election.getLeader())
         .then(currentLeaderKey => {
           expect(phase).to.equal(2);
-          expect(currentLeaderKey).to.equal(election3.leaderKey);
+          expect(currentLeaderKey).to.equal(election3.campaignData!.leaderKey);
         });
 
       // ensure client3 joined campaign
-      await sleep(100);
+      await delay(100);
 
       phase = 1;
 
       await election.resign();
 
       // ensure client2 and client3 watcher triggered
-      await sleep(100);
+      await delay(100);
 
       phase = 2;
 
       await election2.resign();
 
-      await sleep(100);
+      await delay(100);
 
       await election3.resign();
 
       await Promise.all([waitElection2, waitElection3]);
     });
 
-    it('should proclaim if campaign repeatly', async () => {
-      expect(election.isCampaigning).to.be.true;
+    it('should throw if campaign repeatly', async () => {
+      expect(election.campaign).to.be.ok;
 
-      const oldValue = await client.get(election.leaderKey);
+      const oldValue = await client.get(election.campaignData!.leaderKey);
       expect(oldValue).to.equal('candidate');
 
-      await election.campaign('new-value');
-      const newValue = await client.get(election.leaderKey);
-      expect(newValue).to.equal('new-value');
+      expect(() => election.campaign('new-value')).to.throw(InvalidOperationError);
     });
 
-    it('only proclaim if value changed', async () => {
-      const proclaimFn = sinon
-        .stub(election, 'proclaim')
-        .callsFake(Election.prototype.proclaim.bind(election));
+    describe('interruptions', () => {
+      const interruptState = [
+        CampaignState.CreatingLease,
+        CampaignState.CreatingOwnKey,
+        CampaignState.Follower,
+      ];
 
-      await election.campaign('candidate');
-      expect(proclaimFn.notCalled).to.be.true;
-      await election.campaign('candidate2');
-      expect(proclaimFn.calledOnce).to.be.true;
+      const ensureClean = async () => {
+        await delay(50);
+        expect(election.state.getValue().state).to.equal(CampaignState.Idle);
+        expect(await client.getAll().keys()).to.be.empty;
+      };
 
-      proclaimFn.restore();
+      interruptState.forEach(state => {
+        it(`unsubscribes when ${state}`, async () => {
+          await election.resign();
+
+          await election
+            .campaignAdvanced('a')
+            .pipe(takeWhile(s => s.state !== state))
+            .toPromise();
+
+          await ensureClean();
+        });
+      });
+
+      interruptState.forEach(state => {
+        it(`resigns when ${state}`, async () => {
+          await election.resign();
+
+          const resign = new BehaviorSubject(false);
+          let completed = false;
+
+          await election
+            .campaignAdvanced('a')
+            .pipe(takeWhile(s => s.state !== state))
+            .subscribe(
+              s => {
+                if (resign.getValue()) {
+                  throw new Error(`expected to stop after resigning, but got ${s.state}`);
+                }
+
+                if (s.state === state) {
+                  resign.next(true);
+                }
+              },
+              undefined,
+              () => (completed = true),
+            );
+
+          await resign.pipe(take(1)).toPromise();
+          await election.resign();
+          expect(completed).to.be.true;
+          await ensureClean();
+        });
+      });
     });
   });
 
   describe('proclaim', () => {
     it('should update if is a leader', async () => {
-      const oldValue = await client.get(election.leaderKey);
+      const oldValue = await client.get(election.campaignData!.leaderKey);
       expect(oldValue).to.equal('candidate');
       await election.proclaim('new-candidate');
-      const newValue = await client.get(election.leaderKey);
+      const newValue = await client.get(election.campaignData!.leaderKey);
       expect(newValue).to.equal('new-candidate');
     });
 
@@ -122,7 +165,7 @@ describe('election', () => {
   describe('getLeader', () => {
     it('should return leader key', async () => {
       const leaderKey = await election.getLeader();
-      expect(election.leaderKey).to.equal(leaderKey);
+      expect(election.campaignData).to.containSubset({ leaderKey });
     });
 
     it('should throw if no leader', async () => {
@@ -134,46 +177,49 @@ describe('election', () => {
   });
 
   describe('observe', () => {
-    it('should emit leader event', async () => {
-      const client2 = new Etcd3(getOptions());
-      const election2 = new Election(client2, 'test-election', 1);
+    const awaitLeaderValue = (
+      observer: Observable<IKeyValue | undefined>,
+      leader: string | undefined,
+    ) =>
+      observer
+        .pipe(
+          filter(v => (leader ? !!v && v.value.toString() === leader : v === undefined)),
+          first(),
+        )
+        .toPromise();
 
-      let currentLeaderKey = '';
-      election.on('leader', leaderKey => (currentLeaderKey = leaderKey));
+    it('emits when leaders change', async () => {
+      const election2 = new Etcd3(getOptions()).election('test-election', 1);
+      const observer = election.observe().pipe(share());
+      const subscription = observer.subscribe(); // keep the observable hot
 
-      expect(election.isObserving).to.be.true;
-
-      // looking for current leader and emit it
-      await sleep(30);
-
-      expect(currentLeaderKey).to.equal(election.leaderKey);
+      let leader = awaitLeaderValue(observer, 'candidate');
+      await leader;
 
       const waitElection2 = election2.campaign('candidate2');
-
+      leader = awaitLeaderValue(observer, 'candidate2');
       await election.resign();
 
       await waitElection2;
+      await leader;
 
-      // waiting for watcher
-      await sleep(30);
+      election2.resign();
+      await awaitLeaderValue(observer, undefined);
 
-      expect(currentLeaderKey).to.equal(election2.leaderKey);
+      subscription.unsubscribe();
     });
 
-    it('should wait for leader', async () => {
-      await election.resign();
+    it('emits when new values are proclaimed', async () => {
+      const observer = election.observe().pipe(share());
+      const subscription = observer.subscribe(); // keep the observable hot
 
-      let currentLeaderKey = '';
-      election.on('leader', leaderKey => (currentLeaderKey = leaderKey));
+      await awaitLeaderValue(observer, 'candidate');
 
-      // waiting for watcher created
-      await sleep(30);
+      const newValue = awaitLeaderValue(observer, 'new value!');
+      election.proclaim('new value!');
+      await newValue;
 
-      await election.campaign('candidate');
-
-      await sleep(30);
-
-      expect(currentLeaderKey).to.equal(election.leaderKey);
+      subscription.unsubscribe();
     });
   });
 });